Fixes of -Wconversion warnings (not examples).

+++ include/stc/cbits.h
@@ -60,7 +60,7 @@ int main() {
 #define _cbits_bytes(n) (_cbits_words(n) * sizeof(uint64_t))
 
 #if defined(__GNUC__) || defined(__clang__)
-    STC_INLINE uint64_t cpopcount64(uint64_t x) {return __builtin_popcountll(x);}
+    STC_INLINE uint64_t cpopcount64(uint64_t x) {return (uint64_t)__builtin_popcountll(x);}
 #elif defined(_MSC_VER) && defined(_WIN64)
     #include <intrin.h>
     STC_INLINE uint64_t cpopcount64(uint64_t x) {return __popcnt64(x);}
